backend/requirements.txt
absl-py==2.1.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.9.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cryptography==45.0.6
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.0.0
fastapi==0.109.0
fastjsonschema==2.18.1
ffmpeg==1.4
filelock==3.17.0
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.50.0
fqdn==1.5.1
gast==0.5.4
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.62.1
h11==0.16.0
h5py==3.10.0
idna==3.4
imageio==2.34.0
imbalanced-learn==0.13.0
iniconfig==2.1.0
ipykernel==6.25.2
ipython==8.16.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==3.1.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==18.1.1
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matlib==0.1.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mysql-connector-python==9.4.0
namex==0.0.7
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.61.2
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.10.0
pluggy==1.6.0
pocketsphinx==5.0.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.2
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.1.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2024.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.3
rsa==4.9.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
shap==0.47.2
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.25
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.35.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.10.1
types-python-dateutil==2.8.19.14
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.0.5
uvicorn==0.27.0
virtualenv==20.29.1
virtualenvwrapper-win==1.2.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.9
wrapt==1.16.0
xgboost==2.1.4
